+++ ServerDevcommands/Commands/DefaultAutoComplete.cs
@@ -10,20 +10,20 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("challenge");
     AutoComplete.RegisterEmpty("cheers");
     AutoComplete.RegisterEmpty("clear");
-    AutoComplete.Register("fov ", (int index) =>
+    AutoComplete.Register("fov ", index =>
     {
       if (index == 0) return ParameterInfo.Create("Amount", "a positive number");
       return ParameterInfo.None;
     });
     AutoComplete.RegisterEmpty("hidebetatext");
-    AutoComplete.Register("help ", (int index) =>
+    AutoComplete.Register("help ", index =>
     {
       if (index == 0) return ParameterInfo.Create("Page", "number");
       if (index == 1) return ParameterInfo.Create("Page size", "a positive integer (default is 5)");
       return ParameterInfo.None;
     });
     AutoComplete.RegisterEmpty("info");
-    AutoComplete.Register("lodbias", (int index) =>
+    AutoComplete.Register("lodbias", index =>
     {
       if (index == 0) return ParameterInfo.Create("Amount", "a positive number");
       return ParameterInfo.None;
@@ -41,7 +41,7 @@ public static class DefaultAutoComplete
       return options;
     }
     Terminal.commands["raiseskill"].m_tabOptionsFetcher = newFetcher;
-    AutoComplete.Register("raiseskill", (int index) =>
+    AutoComplete.Register("raiseskill", index =>
     {
       var fetcher = Terminal.commands["raiseskill"].m_tabOptionsFetcher;
       if (index == 0) return fetcher();
@@ -51,11 +51,11 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("resetsharedmap");
     AutoComplete.RegisterEmpty("resetspawn");
     AutoComplete.RegisterEmpty("respawn");
-    AutoComplete.Register("s", (int index) =>
+    AutoComplete.Register("s", index =>
     {
       return ParameterInfo.Create("Message");
     });
-    AutoComplete.Register("say", (int index) =>
+    AutoComplete.Register("say", index =>
     {
       return ParameterInfo.Create("Message");
     });
@@ -65,7 +65,7 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("tutorialreset");
     AutoComplete.RegisterEmpty("tutorialtoggle");
     AutoComplete.RegisterEmpty("wave");
-    AutoComplete.Register("W", (int index) =>
+    AutoComplete.Register("W", index =>
     {
       if (index == 0) return ParameterInfo.PlayerNames;
       return ParameterInfo.Create("Message");
@@ -77,7 +77,7 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("save");
     AutoComplete.RegisterAdmin("unban");
 
-    AutoComplete.Register("beard", (int index) =>
+    AutoComplete.Register("beard", index =>
     {
       if (index == 0) return ParameterInfo.Beards;
       return ParameterInfo.None;
@@ -86,14 +86,14 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("dpsdebug");
     AutoComplete.RegisterEmpty("exploremap");
     // Event added to the replaced command.
-    AutoComplete.Register("ffsmooth", (int index) =>
+    AutoComplete.Register("ffsmooth", index =>
     {
       if (index == 0) return ParameterInfo.Create("0 = normal, 1 = add smooth movement");
       return ParameterInfo.None;
     });
     AutoComplete.RegisterEmpty("fly");
     AutoComplete.RegisterEmpty("freefly");
-    AutoComplete.Register("forcedelete", (int index) =>
+    AutoComplete.Register("forcedelete", index =>
     {
       if (index == 0) return ParameterInfo.Create("Radius", "in meters (from 0.0 to 50.0, default is 5.0).");
       if (index == 1) return ParameterInfo.ObjectIds;
@@ -102,13 +102,13 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("gc");
     AutoComplete.RegisterEmpty("ghost");
     AutoComplete.RegisterEmpty("god");
-    AutoComplete.Register("hair", (int index) =>
+    AutoComplete.Register("hair", index =>
     {
       if (index == 0) return ParameterInfo.Hairs;
       return ParameterInfo.None;
     });
     AutoComplete.RegisterEmpty("heal");
-    AutoComplete.Register("itemset", (int index) =>
+    AutoComplete.Register("itemset", index =>
     {
       if (index == 0) return Terminal.commands["itemset"].m_tabOptionsFetcher();
       if (index == 1) return ["keep", "clear"];
@@ -117,25 +117,25 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("killenemycreatures");
     AutoComplete.RegisterEmpty("killtame");
     AutoComplete.RegisterEmpty("listkeys");
-    AutoComplete.Register("location", (int index) =>
+    AutoComplete.Register("location", index =>
     {
       if (index == 0) return ParameterInfo.LocationIds;
       if (index == 1) return ["SAVE"];
       return ParameterInfo.None;
     });
-    AutoComplete.Register("maxfps", (int index) =>
+    AutoComplete.Register("maxfps", index =>
     {
       if (index == 0) return ParameterInfo.Create("Amount", "a positive integer");
       return ParameterInfo.None;
     });
-    AutoComplete.Register("model", (int index) =>
+    AutoComplete.Register("model", index =>
     {
       if (index == 0) return ParameterInfo.Create("<color=yellow>0</color> = male, <color=yellow>1</color> = female");
       return ParameterInfo.None;
     });
     AutoComplete.RegisterEmpty("nocost");
     AutoComplete.RegisterEmpty("noportals");
-    AutoComplete.Register("players", (int index) =>
+    AutoComplete.Register("players", index =>
     {
       if (index == 0) return ParameterInfo.Create("Amount", "a positive integer (0 disables the override)");
       return ParameterInfo.None;
@@ -148,7 +148,7 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("removebirds");
     AutoComplete.RegisterEmpty("removedrops");
     AutoComplete.RegisterEmpty("removefish");
-    AutoComplete.Register("recall ", (int index) =>
+    AutoComplete.Register("recall ", index =>
     {
       if (index == 0) return ParameterInfo.PlayerNames;
       return ParameterInfo.None;
@@ -156,18 +156,18 @@ public static class DefaultAutoComplete
     AutoComplete.RegisterEmpty("resetcharacter");
     AutoComplete.RegisterEmpty("resetenv");
     AutoComplete.RegisterEmpty("resetwind");
-    AutoComplete.Register("removekey", (int index) =>
+    AutoComplete.Register("removekey", index =>
     {
       if (index == 0) return ParameterInfo.GlobalKeys;
       return ParameterInfo.None;
     });
-    AutoComplete.Register("setkey", (int index) =>
+    AutoComplete.Register("setkey", index =>
     {
       if (index == 0) return ParameterInfo.GlobalKeys;
       return ParameterInfo.None;
     });
     AutoComplete.RegisterDefault("setpower");
-    AutoComplete.Register("spawn", (int index) =>
+    AutoComplete.Register("spawn", index =>
     {
       if (index == 0) return ParameterInfo.ObjectIds;
       if (index == 1) return ParameterInfo.Create("Amount", "a positive integer (default 1)");
@@ -176,19 +176,20 @@ public static class DefaultAutoComplete
       return ParameterInfo.None;
     });
     AutoComplete.RegisterEmpty("tame");
-    AutoComplete.Register("test", (int index) =>
+    AutoComplete.Register("test", index =>
     {
       if (index == 0) return ["oldcomfort"];
       return ParameterInfo.None;
     });
+    AutoComplete.RegisterDefault("resetskill");
     AutoComplete.RegisterEmpty("time");
-    AutoComplete.Register("timescale", (int index) =>
+    AutoComplete.Register("timescale", index =>
     {
       if (index == 0) return ParameterInfo.Create("Multiplier", "sets how fast the time goes (from 0.0 to 3.0). Value 0 can be used to pause the game.");
       if (index == 1) return ParameterInfo.Create("Transition duration", "causes the change to be applied gradually over time (seconds). Default value 0 applies the change instantly.");
       return ParameterInfo.None;
     });
-    AutoComplete.Register("tod", (int index) =>
+    AutoComplete.Register("tod", index =>
     {
       if (index == 0) return ParameterInfo.Create("Time", "overrides the time of the day (from 0.0 to 1.0, with 0.5 being the mid day). Value -1 removes the override.");
       return ParameterInfo.None;