+++ ServerDevcommands/ServerDevcommands.cs
@@ -12,7 +12,7 @@ public class ServerDevcommands : BaseUnityPlugin
 {
   public const string GUID = "server_devcommands";
   public const string NAME = "Server Devcommands";
-  public const string VERSION = "1.94";
+  public const string VERSION = "1.94.1";
   public const string COMFY_GIZMO_GUID = "bruce.valheim.comfymods.gizmo";
   public const string RELOADED_GIZMO_GUID = "m3to.mods.GizmoReloaded";
   private static ManualLogSource? Logs;
@@ -113,7 +113,6 @@ public class SetCommands
     new WindCommand();
     new EnvCommand();
     new GotoCommand();
-    new ResetSkillCommand();
     new InventoryCommand();
     new CalmCommand();
     new RepairCommand();