+++ ServerDevcommands/data/BindData.cs
@@ -1,13 +1,33 @@
+using System.Collections.Generic;
 using System.ComponentModel;
+using UnityEngine;
 
 namespace ServerDevcommands;
 
 public class BindData
 {
-  public string key = "";
   [DefaultValue("")]
-  public string modifiers = "";
+  public string keys = "";
+  [DefaultValue("")]
+  public string state = "";
+  public string? key;
+  public string? modifiers;
+  [DefaultValue("")]
   public string command = "";
   [DefaultValue("")]
-  public string tag = "";
+  public string offCommand = "";
+}
+
+public class CommandBind
+{
+  public List<KeyCode> Required = [];
+  public List<KeyCode>? Banned;
+  public List<string>? RequiredState;
+  public List<string>? BannedState;
+  public bool MouseWheel;
+  public string Command = "";
+  public string OffCommand = "";
+  public bool Executed = false;
+  public bool WasExecuted = false;
+  public string? Keys;
 }