+++ publish/CHANGELOG.md
@@ -1,3 +1,8 @@
+- v1.95
+  - Overhauls the bind system to support more features (and simplify implementation).
+  - Removes the setting "Mouse wheel bind key" as obsolete.
+  - Removes the custom `resetskill` command as obsolete.
+
 - v1.94
   - Adds a new setting "Server chat name" to change the name of the server chat.